Cypress shows 3/5 tests pass but all requirements are meet - show attendees for a single event and form HTML format is the same
